pyproject.toml
[project]
name = "pocketwatcher"
version = "0.2.0"
description = "Pocket-sized brute-force & password-spray detector for Linux auth.log and Windows Security logs, with optional GeoIP/ASN detection"
authors = [{name="Brittany Norman"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "rich>=13.7.1",
    "pydantic>=2.7.0",
    "click>=8.1.7",
    "python-dateutil>=2.9.0.post0",
    "ruamel.yaml>=0.18.6",
    "python-evtx>=0.7.4; platform_system == 'Windows'",
]
# Optional extras for enrichment
[project.optional-dependencies]
geo = ["geoip2>=4.8.0"]
asn = ["pyasn>=1.6.1"]

[project.scripts]
pocketwatcher = "pocketwatcher.cli:main"

[tool.ruff]
line-length = 100
target-version = "py39"
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["pocketwatcher"]
force-single-line = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
